# server/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: felixtj.dietrich@tum.de
    name: Felix T.J. Dietrich
  description: API documentation for the Hephaestus application server.
  license:
    name: MIT License
    url: https://github.com/ls1intum/Hephaestus/blob/develop/LICENSE
  title: Hephaestus API
  version: 0.0.0-development
servers:
- description: Default Server URL
  url: /
security:
- bearerAuth: []
...
paths:
...
  /workspaces/{workspaceSlug}/practices:
    get:
      description: "Returns all practice definitions for the workspace, optionally\
        \ filtered by active state"
      operationId: listPractices
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: Filter by active state
        in: query
        name: active
        required: false
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Practice"
          description: Practices returned
      security: []
      summary: List practice definitions
      tags:
      - Practice Catalog
    post:
      operationId: createPractice
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePracticeRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
          description: Practice created
        "409":
          description: Practice slug already exists in this workspace
      summary: Create a new practice definition
      tags:
      - Practice Catalog
...
  /workspaces/{workspaceSlug}/practices/learner:
    get:
      description: "Active practices projected for a developer: name, area, why-it-matters,\
        \ what-good-looks-like. The detection criteria is ABSENT BY CONSTRUCTION (LearnerPracticeDTO\
        \ has no such field)."
      operationId: listLearnerPractices
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearnerPractice"
          description: Learner practices returned
      security: []
      summary: "List active practices, learner-facing"
      tags:
      - Practice Catalog
...
  /workspaces/{workspaceSlug}/practices/reorder:
    patch:
      description: Sets each practice's display order to its index in the provided
        slug list (one atomic write)
      operationId: reorderPractices
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderPracticesRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Practice"
          description: Practices reordered; the full ordered practice list is returned
        "400":
          description: "orderedSlugs is empty, has duplicates, or is not the area's\
            \ complete set"
      summary: Reorder the practices within an area
      tags:
      - Practice Catalog
  /workspaces/{workspaceSlug}/practices/{practiceSlug}:
    delete:
      operationId: deletePractice
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: practiceSlug
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Practice deleted
        "404":
          description: Practice not found
      summary: Delete a practice definition
      tags:
      - Practice Catalog
    get:
      operationId: getPractice
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: practiceSlug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
          description: Practice returned
        "404":
          description: Practice not found
      security: []
      summary: Get a practice definition
      tags:
      - Practice Catalog
    patch:
      operationId: updatePractice
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: practiceSlug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePracticeRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
          description: Practice updated
        "404":
          description: Practice not found
      summary: Update a practice definition
      tags:
      - Practice Catalog
  /workspaces/{workspaceSlug}/practices/{practiceSlug}/active:
    patch:
      operationId: setActive
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: practiceSlug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePracticeActiveRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
          description: Active state updated
        "404":
          description: Practice not found
      summary: Set practice active state
      tags:
      - Practice Catalog
  /workspaces/{workspaceSlug}/practices/{practiceSlug}/area:
    put:
      description: "Binds the practice to the area named by areaSlug, or unbinds it\
        \ when areaSlug is null"
      operationId: bindArea
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: practiceSlug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindPracticeAreaRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
          description: Binding updated
        "404":
          description: Practice or area not found
      summary: Bind a practice to an area
      tags:
      - Practice Catalog
...
components:
  schemas:
...
    BindPracticeAreaRequest:
      type: object
      description: "Request to bind a practice to an area, or unbind it when areaSlug\
        \ is null"
      properties:
        areaSlug:
          type: string
          description: "Slug of the area to bind to, or null to unbind"
          example: review-ready-work
...
    CreatePracticeRequest:
      type: object
      description: Request to create a new practice definition
      properties:
        artifactType:
          type: string
          description: Artifact this practice evaluates. Defaults to PULL_REQUEST
            when omitted.
          enum:
          - PULL_REQUEST
          - ISSUE
          - CONVERSATION_THREAD
          example: PULL_REQUEST
        criteria:
          type: string
          description: Practice evaluation criteria
          maxLength: 50000
          minLength: 0
        name:
          type: string
          description: Human-readable name
          example: PR Description Quality
          maxLength: 128
          minLength: 3
        precomputeScript:
          type: string
          description: TypeScript/Bun precompute script for static analysis before
            AI review
          maxLength: 100000
          minLength: 0
        slug:
          type: string
          description: URL-safe identifier unique within the workspace
          example: pr-description-quality
          maxLength: 64
          minLength: 3
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        triggerEvents:
          type: array
          description: Domain events that trigger detection
          example:
          - PullRequestCreated
          - ReviewSubmitted
          items:
            type: string
          maxItems: 10
          minItems: 1
        whatGoodLooksLike:
          type: string
          description: "Developer-facing exemplar (learner layer); a concrete instance,\
            \ not the rubric"
          maxLength: 2000
          minLength: 0
        whyItMatters:
          type: string
          description: "Developer-facing rationale (learner layer); plain language,\
            \ never the detection rubric"
          maxLength: 2000
          minLength: 0
      required:
      - criteria
      - name
      - slug
      - triggerEvents
...
    LearnerPractice:
      type: object
      description: Learner-facing view of a practice — criteria absent by construction
      properties:
        areaSlug:
          type: string
          description: "Slug of the practice area this belongs to, if any"
        name:
          type: string
          description: Human-readable name
        slug:
          type: string
          description: URL-safe identifier
        whatGoodLooksLike:
          type: string
          description: A concrete picture of doing this well
        whyItMatters:
          type: string
          description: "Why this practice matters, in plain language"
      required:
      - name
      - slug
...
    Practice:
      type: object
      description: Practice definition for evaluating developer contributions
      properties:
        active:
          type: boolean
          description: Whether this practice is actively being detected
        areaSlug:
          type: string
          description: "Slug of the practice area this practice is bound to, if any"
        artifactType:
          type: string
          description: Artifact this practice evaluates
          enum:
          - PULL_REQUEST
          - ISSUE
          - CONVERSATION_THREAD
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the practice was created
        criteria:
          type: string
          description: Practice evaluation criteria
        displayOrder:
          type: integer
          format: int32
          description: Position within its area (lowest first); ties broken by name
        id:
          type: integer
          format: int64
          description: Practice ID
        name:
          type: string
          description: Human-readable name
        precomputeScript:
          type: string
          description: TypeScript/Bun precompute script for static analysis before
            AI review
        slug:
          type: string
          description: URL-safe identifier unique within workspace
        triggerEvents:
          type: array
          description: Domain events that trigger detection
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the practice was last updated
        whatGoodLooksLike:
          type: string
          description: Developer-facing exemplar (learner layer)
        whyItMatters:
          type: string
          description: Developer-facing rationale (learner layer)
      required:
      - active
      - artifactType
      - createdAt
      - criteria
      - displayOrder
      - id
      - name
      - slug
      - triggerEvents
      - updatedAt
...
    ReorderPracticesRequest:
      type: object
      description: Reorder the practices in one area — displayOrder follows the list
        index
      properties:
        areaSlug:
          type:
          - string
          - "null"
          description: Slug of the area whose practices are being reordered; null
            reorders the unassigned bucket
          example: review-ready-work
        orderedSlugs:
          type: array
          description: Practice slugs in the desired display order — the complete
            set for the area
          example:
          - describe-what-and-why
          - ready-and-traceable-handoff
          items:
            type: string
          minItems: 1
      required:
      - orderedSlugs
...
    UpdatePracticeActiveRequest:
      type: object
      description: Request to set a practice's active state
      properties:
        active:
          type: boolean
          description: Whether the practice should be active
      required:
      - active
...
    UpdatePracticeRequest:
      type: object
      description: Request to update an existing practice definition (PATCH — only
        non-null fields applied)
      properties:
        artifactType:
          type: string
          description: Artifact this practice evaluates
          enum:
          - PULL_REQUEST
          - ISSUE
          - CONVERSATION_THREAD
          example: ISSUE
        criteria:
          type: string
          description: Practice evaluation criteria
          maxLength: 50000
          minLength: 0
          pattern: .*\S.*
        name:
          type: string
          description: Human-readable name
          example: PR Description Quality
          maxLength: 128
          minLength: 3
          pattern: .*\S.*
        precomputeScript:
          type: string
          description: TypeScript/Bun precompute script for static analysis before
            AI review
          maxLength: 100000
          minLength: 0
        triggerEvents:
          type: array
          description: Domain events that trigger detection
          items:
            type: string
          maxItems: 10
          minItems: 1
        whatGoodLooksLike:
          type: string
          description: "Developer-facing exemplar (learner layer); a concrete instance,\
            \ not the rubric"
          maxLength: 2000
          minLength: 0
          pattern: .*\S.*
        whyItMatters:
          type: string
          description: "Developer-facing rationale (learner layer); plain language,\
            \ never the detection rubric"
          maxLength: 2000
          minLength: 0
          pattern: .*\S.*
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Hephaestus-native JWT bearer authentication. The SPA normally authenticates
        via the `__Host-HEPHAESTUS_AT` session cookie; this scheme documents the equivalent
        bearer token.
      scheme: bearer
      type: http